requirements.txt
python-dotenv

# Django
django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
whitenoise==6.4.0  # for serving static files

# Django 3.0 async requirements
gunicorn==20.1.0
uvicorn==0.21.1

# Databases
psycopg2==2.9.6
dj-database-url==1.3.0

# Distributed async tasks
celery==5.2.7
redis==4.5.4

# Telegram bot
python-telegram-bot==13.5

# monitoring
# sentry-sdk
